% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDELT.R
\name{CDELT}
\alias{CDELT}
\title{Delta selection}
\usage{
CDELT(dts, m_ad, nbf = 15, delta = seq(from = 0.5, to = 1, by = 0.01))
}
\arguments{
\item{dts}{The matrix with dataset; Number of timepoints X Number of nodes}

\item{m_ad}{Square Matrix Adjacent with dimension, Number of nodes  1 if edge exists; 0 otherwise}

\item{nbf}{the Log Predictive Likelihood will be calculate from this time point. It has to be a positive integer number; The default is 15}

\item{delta}{The vector with the sequence of all discount factors}
}
\description{
Delta selection
}
