% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_filt.R
\name{dlm_filt}
\alias{dlm_filt}
\title{Dynamic linear models (DLMs) filter}
\usage{
dlm_filt(
  Yt,
  Ft,
  delta,
  Gt = array(diag(nrow(Ft)), dim = c(nrow(Ft), nrow(Ft), length(Yt))),
  m0 = rep(0, nrow(Ft)),
  CS0 = 3 * diag(nrow(Ft)),
  n0 = 0.001,
  d0 = 0.001
)
}
\arguments{
\item{Yt}{The vector of observed time series with length T}

\item{Ft}{The matrix of covariates with dimension: number of thetas (p) X sample size (T)}

\item{delta}{Discount factor | Wt=Ctx(1-delta)/delta}

\item{Gt}{The matrix of state equation with dimension: p X p X T. The default is identity matrix block}

\item{m0}{The vector of prior mean at time t=0 with length p. The default is non-informative prior, with zero mean.}

\item{CS0}{The squared matrix of prior variance - C*0 | C*0Vt = C0, with length p. The default is non-informative prior, with prior variance equal to 3 times the observed variance.}

\item{n0}{The prior hypermarameters of precision phi ~ G(n0/2; d0/2). The default is non-informative prior, with value of 0.001. n0 has to be higher than 0.}

\item{d0}{The prior hypermarameters of precision phi ~ G(n0/2; d0/2). The default is non-informative prior, with value of 0.001. n0 has to be higher than 0.}
}
\description{
DLM with Filtering for unknown observational and state variances
}
