% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\name{gower}
\alias{gower}
\title{Compute Gower's centered similarity matrix from a distance matrix}
\usage{
gower(d.mat)
}
\arguments{
\item{d.mat}{Symmetric distance matrix (or R distance object) computed from
the outcome data to be used in MDMR.}
}
\value{
G Gower's centered dissimilarity matrix computed from D.
}
\description{
Compute Gower's centered similarity matrix \eqn{G}, which is the matrix
decomposed by the MDMR test statistic.
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}
\references{
Gower, J. C. (1966). Some distance properties of latent root and
vector methods used in multivariate analysis. Biometrika, 53(3-4), 325-338.
}

