% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm_filt.R
\name{mdm_filt}
\alias{mdm_filt}
\title{Multiregression Dynamic Model (MDM) Filter}
\usage{
mdm_filt(dts, m_ad, DF_hat)
}
\arguments{
\item{dts}{The matrix with dataset; Number of timepoints X Number of nodes}

\item{m_ad}{Square Matrix Adjacent with dimension = Number of nodes # 1 if edge exists; 0 otherwise}

\item{DF_hat}{Vector with delta that maximizes the LPL for each node with dimension = Number of nodes}
}
\description{
MDM with Filtering for unknown observational and state variances
}
