% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm_score.R
\name{mdm_score}
\alias{mdm_score}
\title{Multiregression Dynamic Model (MDM) Score}
\usage{
mdm_score(
  data_input,
  nbf = 15,
  delta = seq(from = 0.5, to = 1, by = 0.01),
  GOLB_print = FALSE,
  subjects_length = 1
)
}
\arguments{
\item{data_input}{The DataFrame that holds the multivariate time series data}

\item{nbf}{Numeric, The Log Predictive Likelihood will be calculate from this time point. It has to be a positive integer number; The default is 15}

\item{delta}{Vector, Sequence of all discount factors. The default is seq(from=0.5, to=1.0, by=0.01)}

\item{GOLB_print}{Logical, if TRUE, save a file that can be used in the GOBNILP optimizer}

\item{subjects_length}{Numeric, how many subjects are contained in dataset}
}
\description{
MDM scores
}
