% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm_smoo.R
\name{mdm_smoo}
\alias{mdm_smoo}
\title{Multiregression Dynamic Model (MDM) snooth}
\usage{
mdm_smoo(mt, Ct, Rt, nt, dt)
}
\arguments{
\item{mt}{List with the matrix of posterior mean with dimension p X T}

\item{Ct}{List with the squared matrix of posterior variance with dimension p X p X T}

\item{Rt}{List with the squared matrix of prior variance with dimension p X p X T}

\item{nt}{List with the vector of prior hypermarameters of precision phi with length T}

\item{dt}{List with the vector of prior hypermarameters of precision phi with length T}
}
\description{
MDM with Smoothing for unknown observational and state variances
}
