% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm_structure.R
\name{mdm_structure}
\alias{mdm_structure}
\title{Learn MDM Structure and Return Adjacency Matrix}
\usage{
mdm_structure(
  data_input,
  method = "hc",
  gobnilp_path = NULL,
  nbf = 15,
  delta = seq(0.5, 1, 0.01),
  subjects_length = 1,
  verbose = TRUE
)
}
\arguments{
\item{data_input}{Data frame. Input multivariate time series (possibly with subjects).}

\item{method}{A \code{character} string indicating the method to use for structure learning. Options are:
\describe{
  \item{"hc"}{Hill-climbing structure learning using the `bnlearn` package.}
  \item{"tabu"}{Tabu search greedy search learning using the `bnlearn` package.}
  \item{"h2pc"}{Hybrid HPC (H2PC) learning using the `bnlearn` package.}
  \item{"mmhc"}{Max-Min Hill Climbing (MMHC) learning using the `bnlearn` package.}
  \item{"rsmax2"}{2-phase Restricted Maximization (RSMAX2) learning using the `bnlearn` package.}
  \item{"ipa"}{Integer Programming Approach using GOBNILP with Jaakkola scoring.}
}}

\item{gobnilp_path}{Character. Path to the compiled GOBNILP binary (required if method = "gobnilp").}

\item{nbf}{Numeric. Starting time point for log-predictive likelihood.}

\item{delta}{Numeric vector. Sequence of discount factors.}

\item{subjects_length}{Numeric. Number of subjects in data.}

\item{verbose}{Logical. Whether to print progress.}
}
\value{
A binary adjacency matrix.
}
\description{
This function computes MDM scores and learns the structure of a dynamic Bayesian network,
returning its adjacency matrix using either `bnlearn::`(hc/tabu/mmhc/h2pc/rsmax2) with a custom 
decomposable or GOBNILP ("ipa") optimizer.
}
