% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\name{print.mdmr}
\alias{print.mdmr}
\title{Print MDMR Object}
\usage{
\method{print}{mdmr}(x, ...)
}
\arguments{
\item{x}{Output from \code{mdmr}}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\item{p-value}{Analytical p-values for the omnibus test and each predictor}
}
\description{
\code{print} method for class \code{mdmr}
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

