% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\name{print.mixed.mdmr}
\alias{print.mixed.mdmr}
\title{Print Mixed MDMR Object}
\usage{
\method{print}{mixed.mdmr}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{mixed.mdmr}}}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\item{p-value}{Analytic p-values for the omnibus test and each predictor}
}
\description{
\code{print} method for class \code{mixed.mdmr}
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

