\name{Regions}
\alias{regioes}
\docType{data}
\title{
COVID-19 new cases in the 5 major brazilian regions.
}
\description{
The dataset comprises 160 rows and 5 columns, corresponding to the different Brazilian regions: North, Northeast, Central-West, South, and Southeast. Each row in the dataset represents an epidemiological week spanning from March 2020 to March 2023.
}
\usage{data("dts_4n")}
\format{
  DataFrame
}
\examples{
data(dts_4n)
}
\keyword{datasets}
