% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BN_pygobnilp.R
\name{run_BN_pygobnilp}
\alias{run_BN_pygobnilp}
\title{Network Structure Learning with pyGOBNILP}
\usage{
run_BN_pygobnilp(score_file = NULL, install_gob = FALSE, palim_nodes = NULL)
}
\arguments{
\item{score_file}{character File path to the local scores file}

\item{install_gob}{logical should the pyGOBNILP be install}

\item{palim_nodes}{numberical number of of nodes (variables).}
}
\description{
A function to learn structure of a Bayesian network (BN) from local scores
}
