% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gobnilp.R
\name{run_gobnilp}
\alias{run_gobnilp}
\title{Run GOBNILP Optimizer and Extract Adjacency Matrix}
\usage{
run_gobnilp(
  scores_path,
  gobnilp_path,
  output_path = tempfile(fileext = ".mat")
)
}
\arguments{
\item{scores_path}{Character string. Path to the `.scores` file (Jaakkola format).}

\item{gobnilp_path}{Character string. Full path to the GOBNILP binary executable.}

\item{output_path}{Optional. Path to the output file for the adjacency matrix. If not provided, a temporary file will be used.}
}
\value{
A binary adjacency matrix representing the learned DAG.
}
\description{
Calls the GOBNILP binary on a precomputed `.scores` file and extracts the resulting 
adjacency matrix from the output.
}
