% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\docType{data}
\name{X.clust}
\alias{X.clust}
\title{Simulated clustered predictor data to illustrate the Mixed-MDMR function}
\format{An object of class \code{data.frame} with 250 rows and 3 columns.}
\usage{
X.clust
}
\description{
See \code{\link{mixed.mdmr}}.
}
\keyword{datasets}
