% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\docType{data}
\name{X.mdmr}
\alias{X.mdmr}
\title{Simulated predictor data to illustrate the MDMR package.}
\format{An object of class \code{matrix} with 500 rows and 3 columns.}
\usage{
X.mdmr
}
\description{
See package vignette by calling \code{vignette('mdmr-vignette')}.
}
\keyword{datasets}

