% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\docType{data}
\name{Y.clust}
\alias{Y.clust}
\title{Simulated clustered outcome data to illustrate the Mixed-MDMR function}
\format{An object of class \code{data.frame} with 250 rows and 12 columns.}
\usage{
Y.clust
}
\description{
See \code{\link{mixed.mdmr}}.
}
\keyword{datasets}

