% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmr.R
\docType{data}
\name{Y.mdmr}
\alias{Y.mdmr}
\title{Simulated outcome data to illustrate the MDMR package.}
\format{An object of class \code{matrix} with 500 rows and 10 columns.}
\usage{
Y.mdmr
}
\description{
See package vignette by calling \code{vignette('mdmr-vignette')}.
}
\keyword{datasets}

